+++ src/components/Checkbox/tests/index.js
@@ -1,5 +1,0 @@
-import { generateSnapshots } from 'helpers';
-
-import Checkbox from '../index';
-
-generateSnapshots(Checkbox);