+++ src/components/Header/tests/index.js
@@ -1,5 +1,0 @@
-import { generateSnapshots } from 'helpers';
-
-import Header from '../index';
-
-generateSnapshots(Header);