+++ src/components/Modal/tests/index.js
@@ -1,5 +1,0 @@
-import { generateSnapshots } from 'helpers';
-
-import Modal from '../index';
-
-generateSnapshots(Modal);