+++ src/components/NavBar/tests/index.js
@@ -1,4 +1,0 @@
-import { generateSnapshots } from 'helpers';
-import Navbar from '../index';
-
-generateSnapshots(Navbar);