+++ src/components/Radio/tests/index.js
@@ -1,5 +1,0 @@
-import { generateSnapshots } from 'helpers';
-
-import Radio from '../index';
-
-generateSnapshots(Radio);