+++ src/elements/Loader/tests/index.js
@@ -1,5 +1,0 @@
-import { generateSnapshots } from 'helpers';
-
-import Loader from '../index';
-
-generateSnapshots(Loader);