+++ src/elements/TextArea/tests/index.js
@@ -1,5 +1,0 @@
-import { generateSnapshots } from 'helpers';
-
-import TextArea from '../index';
-
-generateSnapshots(TextArea);