+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+# [2.0.0-beta.24] - 05-02-2020
+
+### Fix
+
+- Include react and react-dom as externals in webpack config
+
+# [2.0.0-beta.23] - 05-02-2020
+
+### Fix
+
+- Make table body responsive
+
 # [2.0.0-beta.22] - 24-01-2020
 
 ### Fix